requirements.txt
PyYAML>=3.10
gevent>=1.0
greenlet>=0.4.2
six>=1.5.2
thrift>=0.9.1
wsgiref>=0.1.2
